Move wheel extraction into its own module
Use background extraction for local wheels that require hashes and remove the separate streaming wrapper.
