+++ crates/uv-distribution/src/distribution_database.rs
@@ -1,19 +1,16 @@
 use std::cmp::Reverse;
-use std::fmt::Display;
 use std::future::Future;
 use std::io;
 use std::path::Path;
 use std::pin::Pin;
-use std::sync::{Arc, Mutex, PoisonError};
+use std::sync::Arc;
 use std::task::{Context, Poll};
 
-use either::Either;
-use futures::future::{AbortHandle, Aborted};
 use futures::{FutureExt, TryStreamExt};
 use rayon::in_place_scope;
 use rayon::prelude::*;
 use rustc_hash::FxHashMap;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncSeekExt, AsyncWriteExt, ReadBuf};
+use tokio::io::{AsyncRead, AsyncSeekExt, ReadBuf};
 use tokio::sync::Semaphore;
 use tokio_util::compat::FuturesAsyncReadCompatExt;
 use tracing::{Instrument, info_span, instrument, warn};
@@ -30,11 +27,10 @@ use uv_distribution_types::{
     BuildInfo, BuildableSource, BuiltDist, Dist, DistRef, HashPolicy, Hashed, IndexUrl,
     InstalledDist, Name, SourceDist,
 };
-use uv_extract::dirhash::{DirectoryDigest, DirhashTree, HashedFile, UnhashedFile, dirhash_path};
+use uv_extract::dirhash::{DirectoryDigest, HashedFile};
 use uv_extract::hash::Hasher;
-use uv_fs::{LockedFile, PortablePath, write_atomic};
+use uv_fs::{LockedFile, write_atomic};
 use uv_git::{GIT_LFS, GitError};
-use uv_install_wheel::validate_and_heal_record;
 use uv_platform_tags::Tags;
 use uv_preview::PreviewFeature;
 use uv_pypi_types::{HashDigest, HashDigests, PyProjectToml};
@@ -44,6 +40,7 @@ use uv_types::{BuildContext, BuildStack};
 
 use crate::archive::Archive;
 use crate::error::PythonVersion;
+use crate::extracted_wheel::{ExtractedWheel, HashedWheel};
 use crate::hash::http_hash_algorithms;
 use crate::metadata::{ArchiveMetadata, Metadata};
 use crate::source::SourceDistributionBuilder;
@@ -1152,9 +1149,9 @@ impl<'a, Context: BuildContext> DistributionDatabase<'a, Context> {
             let mut hasher = uv_extract::hash::HashReader::new(file, &mut hashers);
 
             // Unzip the wheel to a temporary directory.
-            let mut extracted = ExtractedWheel::extract_streaming(
+            let (temp_dir, mut extracted) = ExtractedWheel::extract_in_background(
                 &mut hasher,
-                temp_dir.path(),
+                temp_dir,
                 self.content_addressed_cache,
             )
             .await
@@ -1288,206 +1285,6 @@ impl<'a, Context: BuildContext> DistributionDatabase<'a, Context> {
     /// Return the [`ManagedClient`] used by this resolver.
     pub fn client(&self) -> &ManagedClient<'a> {
         &self.client
-    }
-}
-
-/// Per-file digests and the hash tree of an extracted wheel.
-struct HashedWheel {
-    files: Vec<HashedFile>,
-    tree: DirhashTree,
-}
-
-/// Files extracted from a wheel, with or without content-addressing metadata.
-enum ExtractedWheel {
-    Unhashed(Vec<UnhashedFile>),
-    Hashed(HashedWheel),
-}
-
-/// Stop the extraction worker before removing its temporary directory.
-///
-/// The worker holds the directory's lock throughout extraction. Dropping the guard cancels active
-/// work and waits for that lock, or takes the directory immediately if the worker is still queued.
-/// Cleanup completes without relying on the async runtime.
-struct ExtractionGuard {
-    abort: AbortHandle,
-    temp_dir: Arc<Mutex<Option<tempfile::TempDir>>>,
-}
-
-impl Drop for ExtractionGuard {
-    fn drop(&mut self) {
-        self.abort.abort();
-        self.temp_dir
-            .lock()
-            .unwrap_or_else(PoisonError::into_inner)
-            .take();
-    }
-}
-
-impl ExtractedWheel {
-    /// Feed the download through a bounded pipe to a single extraction worker.
-    ///
-    /// Filesystem operations run synchronously on the worker, while downloading and hashing remain
-    /// asynchronous. The pipe applies backpressure without buffering the entire wheel.
-    /// Dropping the future cancels extraction and removes its temporary directory synchronously.
-    /// Successful extraction returns ownership of the directory to the caller.
-    ///
-    /// Extraction can leave unread bytes when ZIP validation is disabled. Callers must drain the
-    /// reader before finalizing download hashes.
-    async fn extract_in_background<R>(
-        mut reader: R,
-        temp_dir: tempfile::TempDir,
-        content_addressed: bool,
-    ) -> Result<(tempfile::TempDir, Self), uv_extract::Error>
-    where
-        R: AsyncRead + Unpin,
-    {
-        const READ_BUFFER_SIZE: usize = 128 * 1024;
-        const PIPE_BUFFER_SIZE: usize = 256 * 1024;
-
-        // Allow the download to get ahead while the worker decompresses and writes files.
-        let (sender, receiver) = tokio::io::duplex(PIPE_BUFFER_SIZE);
-        let (abort, registration) = AbortHandle::new_pair();
-        let guard = ExtractionGuard {
-            abort,
-            temp_dir: Arc::new(Mutex::new(Some(temp_dir))),
-        };
-        let temp_dir = Arc::clone(&guard.temp_dir);
-        let mut extraction = tokio::task::spawn_blocking(move || {
-            let temp_dir = temp_dir.lock().unwrap_or_else(PoisonError::into_inner);
-            let Some(temp_dir) = temp_dir.as_ref() else {
-                return Err(uv_extract::Error::Io(io::Error::other(Aborted)));
-            };
-            let extracted = if content_addressed {
-                let (files, tree) = uv_extract::stream::unzip_blocking_and_hash(
-                    receiver,
-                    temp_dir.path(),
-                    registration,
-                )?;
-                Self::Hashed(HashedWheel { files, tree })
-            } else {
-                let files =
-                    uv_extract::stream::unzip_blocking(receiver, temp_dir.path(), registration)?;
-                Self::Unhashed(files)
-            };
-            Ok::<_, uv_extract::Error>(extracted)
-        });
-        let download = async {
-            // Own the write end so EOF, errors and cancellation all close the pipe.
-            let mut sender = sender;
-            let mut buffer = vec![0; READ_BUFFER_SIZE];
-            loop {
-                let read = reader.read(&mut buffer).await?;
-                if read == 0 {
-                    break;
-                }
-                if let Err(err) = sender.write_all(&buffer[..read]).await {
-                    if err.kind() == io::ErrorKind::BrokenPipe {
-                        // The worker either rejected the archive or finished early because ZIP
-                        // validation is disabled. The caller drains the download in the latter case.
-                        break;
-                    }
-                    return Err(err);
-                }
-            }
-            Ok::<_, io::Error>(())
-        };
-        let extraction = tokio::select! {
-            // Prefer a download error over the resulting truncated-ZIP error if both are ready.
-            biased;
-            download = download => {
-                if download.is_err() {
-                    guard.abort.abort();
-                }
-                // Await the worker so the guard does not block this runtime thread on cleanup.
-                let extraction = extraction.await;
-                download.map_err(uv_extract::Error::Io)?;
-                extraction
-            }
-            // Stop reading even if the server stalls after sending an invalid ZIP entry.
-            extraction = &mut extraction => extraction,
-        };
-        let extracted =
-            extraction.map_err(|err| uv_extract::Error::Io(io::Error::other(err)))??;
-        let temp_dir = guard
-            .temp_dir
-            .lock()
-            .unwrap_or_else(PoisonError::into_inner)
-            .take()
-            .ok_or_else(|| uv_extract::Error::Io(io::Error::other(Aborted)))?;
-        Ok((temp_dir, extracted))
-    }
-
-    /// Extract a wheel from a streaming reader, optionally retaining its per-file digests.
-    async fn extract_streaming<R>(
-        reader: R,
-        target: &Path,
-        content_addressed: bool,
-    ) -> Result<Self, uv_extract::Error>
-    where
-        R: AsyncRead + Unpin,
-    {
-        if content_addressed {
-            let (files, tree) = uv_extract::stream::unzip_and_hash(reader, target).await?;
-            Ok(Self::Hashed(HashedWheel { files, tree }))
-        } else {
-            let files = uv_extract::stream::unzip(reader, target).await?;
-            Ok(Self::Unhashed(files))
-        }
-    }
-
-    /// Extract a wheel from a seekable file, optionally retaining its per-file digests.
-    fn extract_seekable(
-        reader: fs_err::File,
-        target: &Path,
-        content_addressed: bool,
-    ) -> Result<Self, uv_extract::Error> {
-        if content_addressed {
-            let (files, tree) = uv_extract::unzip_and_hash(reader, target)?;
-            Ok(Self::Hashed(HashedWheel { files, tree }))
-        } else {
-            let files = uv_extract::unzip(reader, target)?;
-            Ok(Self::Unhashed(files))
-        }
-    }
-
-    /// Return the hashed wheel if content hashing was enabled.
-    fn into_hashed(self) -> Option<HashedWheel> {
-        match self {
-            Self::Unhashed(_) => None,
-            Self::Hashed(wheel) => Some(wheel),
-        }
-    }
-
-    /// Heal the wheel's `RECORD` and keep its hash tree consistent with the repaired contents.
-    fn validate_and_heal_record(&mut self, root: &Path, dist: impl Display) -> Result<(), Error> {
-        let files = match self {
-            Self::Unhashed(files) => {
-                Either::Left(files.iter().map(|file| (file.path(), file.size())))
-            }
-            Self::Hashed(wheel) => {
-                Either::Right(wheel.files.iter().map(|file| (file.path(), file.size())))
-            }
-        };
-        let Some(record_path) =
-            validate_and_heal_record(root, files, dist).map_err(Error::InstallWheelError)?
-        else {
-            return Ok(());
-        };
-        let Self::Hashed(hashed_wheel) = self else {
-            return Ok(());
-        };
-
-        let hash = dirhash_path(&root.join(&record_path)).map_err(|err| {
-            Error::Extract(
-                record_path.display().to_string(),
-                uv_extract::Error::from(err),
-            )
-        })?;
-        let record_path = PortablePath::from(record_path.as_path()).to_string();
-        hashed_wheel
-            .tree
-            .update_file(&record_path, hash)
-            .map_err(|err| Error::Extract(record_path, uv_extract::Error::from(err)))
     }
 }
