+++ crates/uv-distribution/src/lib.rs
@@ -14,6 +14,7 @@ mod archive;
 mod distribution_database;
 mod download;
 mod error;
+mod extracted_wheel;
 mod hash;
 mod index;
 mod metadata;
